pyproject.toml
[tool.poetry]
name = "advent-of-code-2021"
version = "0.1.0"
description = ""
authors = ["Andy Freeland <andy@andyfreeland.net>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.0.3"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
